The aim is to classify iris flowers among three species (setosa, versicolor, or virginica) from measurements of sepals and petals' length and  width.  The iris data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant.  The central goal here is to design a model that makes useful classifications for new flowers or, in other words, one which exhibits good  generalization.  The data source is the file iris_flowers.csv. It contains the data for this example in comma-separated values (CSV) format. The number of  columns is 5, and the number of rows is 150.
The aim is to classify iris flowers among three species (setosa, versicolor, or virginica) from measurements of sepals and petals' length and width.
The iris data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant.
The central goal here is to design a model that makes useful classifications for new flowers or, in other words, one which exhibits good generalization.
The data source is the file iris_flowers.csv. It contains the data for this example in comma-separated values (CSV) format. The number of
columns is 5, and the number of rows is 150.